Coordinate Space Refactor + Separate Physics World Per Scene (#10441)
* standardize physics world argument across physics api

* Rename SceneComponent “children” field to “scenes”

* multi-scene physics refactor, scene loading and avatar locomotion works

* physics scene ref space fixed

* fix physics world scale, various improvements

* refactor scene component to be simpler

* physics fixes

* fix updating rigidbody parent matrices overriding physics for sleeping entities

* cleanup

* physics raycast account for scene transform

* cleanup log

* bug fix useEngineCanvas

* fix WebXR

* disable xr avatar sync

* new abstraction for reactively handling search params, make spectate & avatar spawning handling data driven

* fix true string to empty string

* add vitest, patch node for webxr APIs

* license

* fix teleport visibility

* webxr reference space refactor, locomotion works well

* fix movement when placed

* add test widget to toggle embodied and dollhouse avatar., fix detected planes and meshes

* remove xr control state

* testing

* cleanup

* add useReactiveRef

* license

* tests use new APIs

* physics tests pass

* replace useScene with useRendererEntity

* license

* fix render test

* fix engine and editor tests

* tsc error

* fix

* eslint

* various fixes when leaving XR

* scene kill floor system fix

* fix bug with collider treansform updates

* fix CSM

* fixes to mock renderer

* better spatial engine mocking

* license

* bug fixes

* add back getDAG and fix avatar jitter

* initial commit

* revert

* fix project tests

* bug fixes

* fix rigidbody transform problem

* disable select root entity in studio

* fix test

* fix gltf state test

* fix editor tests

* fix instanceserver not loading scene

+++ packages/ui/src/components/editor/panels/Hierarchy/container/index.tsx
@@ -252,6 +252,7 @@ function HierarchyPanelContents(props: { sceneURL: string; rootEntityUUID: Entit
         // Exit click placement mode when anything in the hierarchy is selected
         getMutableState(EditorHelperState).placementMode.set(PlacementMode.DRAG)
         if (e.ctrlKey) {
+          if (entity === rootEntity) return
           EditorControlFunctions.toggleSelection([getComponent(entity, UUIDComponent)])
         } else if (e.shiftKey && prevClickedNode) {
           const startIndex = entityHierarchy.value.findIndex((n) => n.entity === prevClickedNode)
@@ -290,6 +291,7 @@ function HierarchyPanelContents(props: { sceneURL: string; rootEntityUUID: Entit
       switch (e.key) {
         case 'ArrowDown': {
           e.preventDefault()
+          if (entity === rootEntity) return
 
           const nextNode = nodeIndex !== -1 && entityHierarchy.value[nodeIndex + 1]
           if (!nextNode) return
@@ -307,6 +309,7 @@ function HierarchyPanelContents(props: { sceneURL: string; rootEntityUUID: Entit
 
         case 'ArrowUp': {
           e.preventDefault()
+          if (entity === rootEntity) return
 
           const prevNode = nodeIndex !== -1 && entityHierarchy.value[nodeIndex - 1]
           if (!prevNode) return
@@ -337,6 +340,7 @@ function HierarchyPanelContents(props: { sceneURL: string; rootEntityUUID: Entit
           break
 
         case 'Enter':
+          if (entity === rootEntity) return
           if (e.shiftKey) {
             EditorControlFunctions.toggleSelection([getComponent(entity, UUIDComponent)])
           } else {
@@ -346,6 +350,7 @@ function HierarchyPanelContents(props: { sceneURL: string; rootEntityUUID: Entit
 
         case 'Delete':
         case 'Backspace':
+          if (entity === rootEntity) return
           if (selectedNodes && !renamingNode) onDeleteNode(entity)
           break
       }
